Prepare deployment on Heroku
- I do like my ops things to be nicely automated and Heroku provide a
  very good way of doing so. The database patches needed fix as well and
  the "hardest" thing to configure was the Postgresql driver. Heroku
  doesn't provide a JDBC formated URL by default and I had to add a
  special Env Variable up there.

+++ conf/evolutions/default/1.sql
@@ -6,9 +6,9 @@ CREATE TABLE invoice (
     id UUID NOT NULL,
     movie_id BIGINT NOT NULL,
     date TIMESTAMP NOT NULL,
-    paid DOUBLE NOT NULL,
+    paid FLOAT NOT NULL,
     return_date TIMESTAMP,
-    extra_charge DOUBLE,
+    extra_charge FLOAT,
     PRIMARY KEY (id)
 );
